# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MenuAI – AI-powered menu translation and allergen platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List all restaurants
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get a restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Not found
    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantBody"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    delete:
      operationId: deleteRestaurant
      tags: [restaurants]
      summary: Delete a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Menus ────────────────────────────────────────────────────
...
  /restaurants/{id}/analytics/summary:
    get:
      operationId: getRestaurantAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard summary for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /restaurants/{id}/analytics/popular-dishes:
    get:
      operationId: getPopularDishes
      tags: [analytics]
      summary: Get most-viewed dishes from QR menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Popular dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PopularDish"
  /restaurants/{id}/specials:
    get:
      operationId: getTonightSpecials
      tags: [restaurants]
      summary: Get tonight's specials for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tonight's specials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"
    patch:
      operationId: updateTonightSpecials
      tags: [restaurants]
      summary: Set tonight's specials (mark dishes as specials)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpecialsBody"
      responses:
        "200":
          description: Updated specials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dish"

  # ── Public digital menu (QR code endpoint) ────────────────────
...
components:
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        cuisineType:
          type: ["string", "null"]
        tier:
          type: string
          enum: [standard, hotel]
        primaryLanguage:
          type: string
          default: "en"
        supportedLanguages:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, slug, tier, primaryLanguage, supportedLanguages, createdAt, updatedAt]
    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        cuisineType:
          type: ["string", "null"]
        tier:
          type: string
          enum: [standard, hotel]
        primaryLanguage:
          type: string
          default: "en"
        supportedLanguages:
          type: array
          items:
            type: string
      required: [name]
    UpdateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        neighborhood:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        cuisineType:
          type: ["string", "null"]
        tier:
          type: string
          enum: [standard, hotel]
        primaryLanguage:
          type: string
        supportedLanguages:
          type: array
          items:
            type: string

    # ── Menu ───────────────────────────────────────────────────
...
    Dish:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
          default: "KES"
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
          default: true
        isSpecial:
          type: boolean
          default: false
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sectionId, name, currency, isAvailable, isSpecial, sortOrder, createdAt, updatedAt]
...
    AnalyticsSummary:
      type: object
      properties:
        totalDishes:
          type: integer
        availableDishes:
          type: integer
        unavailableDishes:
          type: integer
        translatedDishes:
          type: integer
        allergensConfirmed:
          type: integer
        allergensPendingReview:
          type: integer
        totalMenuViews:
          type: integer
        menus:
          type: integer
        specialsCount:
          type: integer
        languagesCovered:
          type: array
          items:
            type: string
      required: [totalDishes, availableDishes, unavailableDishes, translatedDishes, allergensConfirmed, allergensPendingReview, totalMenuViews, menus, specialsCount, languagesCovered]
    PopularDish:
      type: object
      properties:
        dishId:
          type: integer
        name:
          type: string
        sectionName:
          type: string
        viewCount:
          type: integer
        price:
          type: ["number", "null"]
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
      required: [dishId, name, sectionName, viewCount, currency]
    UpdateSpecialsBody:
      type: object
      properties:
        dishIds:
          type: array
          items:
            type: integer
      required: [dishIds]

    # ── Public Menu ────────────────────────────────────────────
...